% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{merge_object}
\alias{merge_object}
\title{Merge redundant objects into one}
\usage{
merge_object(idf, class)
}
\arguments{
\item{idf}{An \link[eplusr:Idf]{eplusr::Idf} object}

\item{class}{A character vector of class names in current \link[eplusr:Idf]{eplusr::Idf}. If
\code{NULL}, all classes in current \link[eplusr:Idf]{eplusr::Idf} object will be used.}
}
\value{
The modified object itself, invisibly
}
\description{
Merge redundant objects into one
}
\details{
Redundant objects here refer to objects that have different names but all
other field values are the same.

\code{merge_object()} will only keep the first object and remove all redundant
objects. The field value references are handled automatically, i.e. all
field values that originally refer to the redundant objects will be
redirected to the object that is kept.
}
\examples{
\dontrun{
merge_object(idf, "ScheduleTypeLimits")
}
}
