% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{purge_object}
\alias{purge_object}
\title{Purge objects that are not referenced by any others}
\usage{
purge_object(idf, class)
}
\arguments{
\item{idf}{An \link[eplusr:Idf]{eplusr::Idf}}

\item{class}{A character vector of classes where objects to purge}
}
\value{
The modified \link[eplusr:Idf]{eplusr::Idf} object itself, invisiblly
}
\description{
Purge objects that are not referenced by any others
}
\examples{
\dontrun{
purge_object(idf, "ScheduleTypeLimits")
}

}
