% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_per_field}
\alias{rename_per_field}
\title{Rename object in specified class based on zone or zone list it belongs to}
\usage{
rename_per_field(
  idf,
  class,
  field,
  class_fun = function(x) abbreviate(x, 10L),
  field_fun = NULL,
  strict = FALSE
)
}
\arguments{
\item{idf}{An \link[eplusr:Idf]{eplusr::Idf}}

\item{class}{A character vector of classes names}

\item{field}{An integer or a single string specifying the fields whose values
will be used as input passed to \code{field_fun}}

\item{class_fun, field_fun}{A function to take the corresponding class names
or field values as input, process them and return a character vector with the
same length. It can also be a named character vector. In this case, when
specified class names or field values match element value, the element name
will be used as part of the new name. If \code{NULL}, the class names or field
values will be directly used. For \code{class_fun}, default is to
use \code{\link[base:abbreviate]{base::abbreviate()}} to abbreviate them into 10 character width. For
\code{field_fun}, default is \code{NULL}.}

\item{strict}{If \code{TRUE}, an error will be issued if the newly generated
object names contain duplications. If \code{FALSE}, \link{make.unique} will be called
to make sure there are no duplications. Default: \code{FALSE}}
}
\value{
The modified \link[eplusr:Idf]{eplusr::Idf} object itself, invisiblly
}
\description{
\code{rename_per_field()} renames all objects in specified class in format
\code{class-field}, where \code{class} and \code{field} are character vectors created by
using class names and \code{class_fun}, and field values and \code{field_fun},
respectivly.
}
