% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_per_fun}
\alias{rename_per_fun}
\title{Rename objects using speicified function}
\usage{
rename_per_fun(idf, class, fun = toCamelCase, strict = FALSE)
}
\arguments{
\item{idf}{An \link[eplusr:Idf]{eplusr::Idf} object}

\item{class}{A character vector of classes names}

\item{fun}{A function to take the corresponding object names as input,
process them and return a character vector with the same length. \code{fun}
can be a named character vector. In this case, when specified object names
match element value, the element name will be used as the new name. If
\code{NULL}, nothing will be done. Default: \link{toCamelCase}}

\item{strict}{If \code{TRUE}, an error will be issued if the newly generated
object names contain duplications. If \code{FALSE}, \link{make.unique} will be called
to make sure there are no duplications. Default: \code{FALSE}}
}
\value{
The modified \link[eplusr:Idf]{eplusr::Idf} object, invisibly
}
\description{
Rename objects using speicified function
}
