% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule_year_to_compact}
\alias{schedule_year_to_compact}
\title{Convert "Schedule:Year" Objects to "Schedule:Compact" objects}
\usage{
schedule_year_to_compact(idf, object = NULL)
}
\arguments{
\item{idf}{An \link[eplusr:Idf]{eplusr::Idf} object}

\item{object}{An integer vector of IDs or a character vector of names of
objects in \code{Schedule:Year} to be converted}
}
\value{
The modifed \code{Idf} object, indivisibly
}
\description{
Convert "Schedule:Year" Objects to "Schedule:Compact" objects
}
\examples{
\dontrun{
schedule_year_to_compact(idf)
}
}
