% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{toCamelCase}
\alias{toCamelCase}
\title{Change to CamelCase}
\usage{
toCamelCase(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector changed to CamelCase
}
\description{
Change to CamelCase
}
\examples{
toCamelCase(c("a b", "a b-c_d"))
}
