% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{verify_class}
\alias{verify_class}
\title{Verify input class names and return only existings}
\usage{
verify_class(idf, class)
}
\arguments{
\item{idf}{An \link[eplusr:Idf]{eplusr::Idf} object}

\item{class}{A character vector of class names to verify}
}
\value{
A character vector
}
\description{
Verify input class names and return only existings
}
