% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{verify_group}
\alias{verify_group}
\title{Verify input group names and return only existings}
\usage{
verify_group(idf, group)
}
\arguments{
\item{idf}{An \link[eplusr:Idf]{eplusr::Idf} object}

\item{group}{A character vector of group names to verify}
}
\value{
A character vector
}
\description{
Verify input group names and return only existings
}
