% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_check_disabled}
\alias{with_check_disabled}
\title{Evaluate an expression with eplusr checking components disabled}
\usage{
with_check_disabled(disable, expr)
}
\arguments{
\item{disable}{Names of checking components to disable during evaluation of
\code{expr}.}

\item{expr}{An expression to be evaluated with specified checking components
disabled.}
}
\description{
Evaluate an expression with eplusr checking components disabled
}
