% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_eplusr_option}
\alias{with_eplusr_option}
\title{Evaluate an expression with temporary eplusr options}
\usage{
with_eplusr_option(opts, expr)
}
\arguments{
\item{opts}{A list of valid input for \code{eplusr::eplusr_option()}.}

\item{expr}{An expression to be evaluated.}
}
\description{
Evaluate an expression with temporary eplusr options
}
